+++ packages/react-widgets/src/widgets/GeocoderWidget.js
@@ -53,12 +53,17 @@ const SearchIcon = (args) => (
  * @param  {object} props
  * @param  {Object} [props.className] - Material-UI withStyle class for styling
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
+ * @param  {Object=} [props.credentials] - Credentials to use, otherwise widgets checks redus state.
+ * @param  {string=} props.method - method, either `sql` or `lds`, see geocodeStreetPoint
+ * @param  {Boolean=} [props.zoomControl] - Optional, default true. Whether control should zoom map on result.
+ *
+ * @see geocodeStreetPoint
  */
 function GeocoderWidget(props) {
   const inputRef = useRef();
   const oauthCredentials = useSelector(selectOAuthCredentials);
   const globalCredentials = useSelector((state) => state.carto.credentials);
-  const credentials = oauthCredentials || globalCredentials;
+  const credentials = props.credentials || oauthCredentials || globalCredentials;
   // Component local state and events handling
   const [searchText, setSearchText] = useState('');
   const [loading, setIsLoading] = useState(false);
@@ -106,10 +111,13 @@ function GeocoderWidget(props) {
         setIsLoading(true);
         const result = await geocodeStreetPoint(credentials, {
           searchText,
+          method: props.method,
           country: DEFAULT_COUNTRY
         });
         if (result) {
-          zoomToResult(result);
+          if (props.zoomControl !== false) {
+            zoomToResult(result);
+          }
           updateMarker(result);
         }
       } catch (e) {