+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- GeocoderWidget support for LDS queries [#387](https://github.com/CartoDB/carto-react/pull/387)
+
 ## 1.3
 
 ### 1.3.0-alpha.6 (2022-04-27)